+++ odyssey-app/src/main/java/com/codebutler/odyssey/app/OdysseyApplicationComponent.kt
@@ -34,9 +34,9 @@ interface OdysseyApplicationComponent {
 
     fun gameLibrary(): GameLibrary
 
-    fun odysseyDb(): OdysseyDatabase
-
     fun gameLibraryProviderRegistry(): GameLibraryProviderRegistry
+
+    fun odysseyDb(): OdysseyDatabase
 
     fun inject(app: OdysseyApplication)
